refactor: move checkbox and radiogroup into form
