+++ packages/clay-card/stories/index.tsx
@@ -6,7 +6,7 @@
 
 import '@clayui/css/lib/css/atlas.css';
 import ClayCard from '../src';
-import ClayCheckbox from '@clayui/checkbox';
+import ClayForm from '@clayui/form';
 import ClayIcon from '@clayui/icon';
 import ClayLabel from '@clayui/label';
 import ClaySticker from '@clayui/sticker';
@@ -22,14 +22,14 @@ const ClayCheckboxWithState = (props: any) => {
 	const [value, setValue] = React.useState<boolean>(false);
 
 	return (
-		<ClayCheckbox
+		<ClayForm.Checkbox
 			checked={value}
 			disabled={false}
 			indeterminate={false}
 			onChange={() => setValue(val => !val)}
 		>
 			{props.children}
-		</ClayCheckbox>
+		</ClayForm.Checkbox>
 	);
 };
