--- packages/clay-checkbox/stories/index.tsx
@@ -1,45 +1,0 @@
-/**
- * © 2019 Liferay, Inc. <https://liferay.com>
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-import '@clayui/css/lib/css/atlas.css';
-import ClayCheckbox from '../src';
-import React, {useState} from 'react';
-import {boolean, text} from '@storybook/addon-knobs';
-import {storiesOf} from '@storybook/react';
-
-const ClayCheckboxWithState = () => {
-	const [value, setValue] = useState<boolean>(false);
-
-	return (
-		<ClayCheckbox
-			checked={value}
-			disabled={boolean('Disabled', false)}
-			indeterminate={boolean('Indeterminate', false)}
-			label={text('Label', 'Default')}
-			onChange={() => setValue(val => !val)}
-		/>
-	);
-};
-
-storiesOf('ClayCheckbox', module)
-	.add('default', () => <ClayCheckboxWithState />)
-	.add('hidden label w/ aria-label', () => (
-		<ClayCheckbox
-			aria-label="hello! Can you see me?"
-			checked={boolean('Checked', false)}
-			onChange={() => {}}
-		/>
-	))
-	.add('custom JSX content', () => (
-		<ClayCheckbox
-			checked={boolean('Checked', false)}
-			label="Badge"
-			onChange={() => {}}
-		>
-			<span className="badge badge-primary">
-				<span className="badge-item badge-item-expand">{'10'}</span>
-			</span>
-		</ClayCheckbox>
-	));
