+++ packages/clay-drop-down/package.json
@@ -27,9 +27,8 @@
 	],
 	"dependencies": {
 		"@clayui/button": "^3.0.0-milestone.1",
-		"@clayui/checkbox": "^3.0.0-milestone.1",
+		"@clayui/form": "^3.0.0-milestone.1",
 		"@clayui/icon": "^3.0.0-milestone.1",
-		"@clayui/radio-group": "^3.0.0-milestone.1",
 		"@clayui/shared": "^3.0.0-milestone.1",
 		"classnames": "^2.2.6",
 		"metal-position": "^2.1.2"
