+++ packages/clay-list/package.json
@@ -30,7 +30,7 @@
 		"classnames": "^2.2.6"
 	},
 	"devDependencies": {
-		"@clayui/checkbox": "^3.0.0-milestone.1",
+		"@clayui/form": "^3.0.0-milestone.1",
 		"@clayui/sticker": "^3.0.0-milestone.1",
 		"@clayui/css": "^3.0.0-milestone.1"
 	},
