--- packages/clay-radio-group/package.json
@@ -1,41 +1,0 @@
-{
-	"name": "@clayui/radio-group",
-	"version": "3.0.0-milestone.1",
-	"description": "ClayRadioGroup component",
-	"license": "BSD-3-Clause",
-	"repository": "https://github.com/liferay/clay/tree/master/packages/clay-radio-group",
-	"engines": {
-		"node": ">=0.12.0",
-		"npm": ">=3.0.0"
-	},
-	"main": "lib/index.js",
-	"types": "lib/index.d.ts",
-	"ts:main": "src/index.tsx",
-	"files": [
-		"lib",
-		"src"
-	],
-	"scripts": {
-		"build": "babel src --root-mode upward --out-dir lib --extensions .ts,.tsx",
-		"build:types": "tsc --project ./tsconfig.declarations.json",
-		"prepublishOnly": "npm run build && npm run build:types",
-		"test": "jest --config ../../jest.config.js"
-	},
-	"keywords": [
-		"clay",
-		"react"
-	],
-	"dependencies": {
-		"classnames": "^2.2.6"
-	},
-	"devDependencies": {
-		"@clayui/css": "^3.0.0-milestone.1"
-	},
-	"peerDependencies": {
-		"react": "^16.8.1",
-		"react-dom": "^16.8.1"
-	},
-	"browserslist": [
-		"extends browserslist-config-clay"
-	]
-}
