--- packages/clay-radio-group/tsconfig.declarations.json
@@ -1,7 +1,0 @@
-{
-	"compilerOptions": {
-		"declarationDir": "./lib"
-	},
-	"extends": "../../tsconfig.declarations.json",
-	"include": ["src"]
-}
