--- packages/clay-radio-group/tsconfig.json
@@ -1,4 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"include": ["src"]
-}
